chore: add entity/model representation from product table & revise price data type on request and response

+++ internal/server/response.go
@@ -42,15 +42,15 @@ type RegisterResponse struct {
 }
 
 type PostProductResponse struct {
-	ProductID        string `json:"productId"`
-	Name             string `json:"name"`
-	Category         string `json:"category"`
-	Qty              int    `json:"qty"`
-	Price            int    `json:"price"`
-	Sku              string `json:"sku"`
-	FileID           string `json:"fileId"`
-	FileUri          string `json:"fileUri"`
-	FileThumbnailUri string `json:"fileThumbnailUri"`
-	CreatedAt        string `json:"createdAt"`
-	UpdatedAt        string `json:"updatedAt"`
+	ProductID        string  `json:"productId"`
+	Name             string  `json:"name"`
+	Category         string  `json:"category"`
+	Qty              int     `json:"qty"`
+	Price            float64 `json:"price"`
+	Sku              string  `json:"sku"`
+	FileID           string  `json:"fileId"`
+	FileUri          string  `json:"fileUri"`
+	FileThumbnailUri string  `json:"fileThumbnailUri"`
+	CreatedAt        string  `json:"createdAt"`
+	UpdatedAt        string  `json:"updatedAt"`
 }
